agent host otel: split IAgentHostOTelService into common/ to fix renderer test loading
Move the interface + decorator to src/vs/platform/agentHost/common/otel/agentHostOTelService.ts. The concrete AgentHostOTelService class stays in node/ alongside its OTLP receiver / SQLite store / outbound forwarder dependencies.

Mirror TelemetryConfig as IAgentHostSdkTelemetryConfig locally to keep the common file free of node-only package imports (@github/copilot-sdk is restricted to hasNode layers).

Unblocks copilotAgent.test in the Electron renderer harness, which previously failed to dynamically import because outboundForwarder.ts statically references url/fs through the old import chain.
